CRC: Move signer_state tests to a dedicated test module
