+++ stacks-signer/src/tests/mod.rs
@@ -1,1 +1,2 @@
 mod chainstate;
+mod signer_state;
